Refactor and clean up search
